requirements.txt
pyTelegramBotAPI==4.21.0
aiohttp==3.9.5
asyncpg==0.29.0
python-dotenv==1.0.1
web3==6.18.0
eth_account>=0.10.0

# Тестирование
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0

# AI зависимости (опционально)
google-genai>=0.3.0
